requirements.txt
pandas
numpy
tqdm
sqlalchemy
sentence-transformers
python-dotenv
pyodbc
